remove function to its own file / namespace
