+++ Source/Make.WarpX
@@ -70,6 +70,7 @@ endif
 
 -include Make.package
 include $(WARPX_HOME)/Source/Make.package
+include $(WARPX_HOME)/Source/ABLASTR/Make.package
 include $(WARPX_HOME)/Source/BoundaryConditions/Make.package
 include $(WARPX_HOME)/Source/Diagnostics/Make.package
 include $(WARPX_HOME)/Source/EmbeddedBoundary/Make.package
