+++ Source/Particles/Deposition/ChargeDeposition.H
@@ -214,126 +214,4 @@ void doChargeDepositionShapeN (const GetParticlePosition& GetPosition,
 #endif
 }
 
-template <typename PC>
-void DepositChargeDoIt (typename PC::ParIterType& pti,
-                        typename PC::RealVector& wp,
-                        const int * const ion_lev,
-                        amrex::MultiFab* rho, const int icomp,
-                        const long offset, const long np_to_depose,
-                        amrex::FArrayBox& local_rho, const int lev, const int depos_lev,
-                        const amrex::Real charge, const int nox, const int noy, const int noz,
-                        const amrex::IntVect& ng_rho, const std::array<amrex::Real,3>& dx,
-                        const amrex::IntVect& ref_ratio, const int nc,
-                        const std::array<amrex::Real, 3>& xyzmin,
-                        amrex::Real* cost, const int n_rz_azimuthal_modes,
-                        const long load_balance_costs_update_algo)
-{
-    AMREX_ALWAYS_ASSERT_WITH_MESSAGE((depos_lev==(lev-1)) ||
-                                     (depos_lev==(lev  )),
-                                     "Deposition buffers only work for lev-1");
-
-    // If no particles, do not do anything
-    if (np_to_depose == 0) return;
-
-    // Extract deposition order and check that particles shape fits within the guard cells.
-    // NOTE: In specific situations where the staggering of rho and the charge deposition algorithm
-    // are not trivial, this check might be too strict and we might need to relax it, as currently
-    // done for the current deposition.
-
-#if   defined(WARPX_DIM_1D_Z)
-    amrex::ignore_unused(nox);
-    amrex::ignore_unused(noy);
-    const amrex::IntVect shape_extent = amrex::IntVect(static_cast<int>(noz/2+1));
-#elif   defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-    amrex::ignore_unused(noy);
-    const amrex::IntVect shape_extent = amrex::IntVect(static_cast<int>(nox/2+1),
-                                                       static_cast<int>(noz/2+1));
-#elif defined(WARPX_DIM_3D)
-    const amrex::IntVect shape_extent = amrex::IntVect(static_cast<int>(nox/2+1),
-                                                       static_cast<int>(noy/2+1),
-                                                       static_cast<int>(noz/2+1));
-#endif
-
-    // On CPU: particles deposit on tile arrays, which have a small number of guard cells ng_rho
-    // On GPU: particles deposit directly on the rho array, which usually have a larger number of guard cells
-#ifndef AMREX_USE_GPU
-    const amrex::IntVect range = ng_rho - shape_extent;
-#else
-    const amrex::IntVect range = rho->nGrowVect() - shape_extent;
-#endif
-
-    AMREX_ALWAYS_ASSERT_WITH_MESSAGE(
-        amrex::numParticlesOutOfRange(pti, range) == 0,
-        "Particles shape does not fit within tile (CPU) or guard cells (GPU) used for charge deposition");
-
-    WARPX_PROFILE_VAR_NS("WarpXParticleContainer::DepositCharge::ChargeDeposition", blp_ppc_chd);
-    WARPX_PROFILE_VAR_NS("WarpXParticleContainer::DepositCharge::Accumulate", blp_accumulate);
-
-    // Get tile box where charge is deposited.
-    // The tile box is different when depositing in the buffers (depos_lev<lev)
-    // or when depositing inside the level (depos_lev=lev)
-    amrex::Box tilebox;
-    if (lev == depos_lev) {
-        tilebox = pti.tilebox();
-    } else {
-        tilebox = amrex::coarsen(pti.tilebox(),ref_ratio);
-    }
-
-#ifndef AMREX_USE_GPU
-    // Staggered tile box
-    amrex::Box tb = amrex::convert( tilebox, rho->ixType().toIntVect() );
-#endif
-
-    tilebox.grow(ng_rho);
-
-#ifdef AMREX_USE_GPU
-    amrex::ignore_unused(local_rho);
-    // GPU, no tiling: rho_fab points to the full rho array
-    amrex::MultiFab rhoi(*rho, amrex::make_alias, icomp*nc, nc);
-    auto & rho_fab = rhoi.get(pti);
-#else
-    tb.grow(ng_rho);
-
-    // CPU, tiling: rho_fab points to local_rho
-    local_rho.resize(tb, nc);
-
-    // local_rho is set to zero
-    local_rho.setVal(0.0);
-
-    auto & rho_fab = local_rho;
-#endif
-
-    const auto GetPosition = GetParticlePosition(pti, offset);
-
-    // Indices of the lower bound
-    const amrex::Dim3 lo = lbound(tilebox);
-
-    WARPX_PROFILE_VAR_START(blp_ppc_chd);
-
-    if        (nox == 1){
-        doChargeDepositionShapeN<1>(GetPosition, wp.dataPtr()+offset, ion_lev,
-                                    rho_fab, np_to_depose, dx, xyzmin, lo, charge,
-                                    n_rz_azimuthal_modes, cost,
-                                    load_balance_costs_update_algo);
-    } else if (nox == 2){
-        doChargeDepositionShapeN<2>(GetPosition, wp.dataPtr()+offset, ion_lev,
-                                    rho_fab, np_to_depose, dx, xyzmin, lo, charge,
-                                    n_rz_azimuthal_modes, cost,
-                                    load_balance_costs_update_algo);
-    } else if (nox == 3){
-        doChargeDepositionShapeN<3>(GetPosition, wp.dataPtr()+offset, ion_lev,
-                                    rho_fab, np_to_depose, dx, xyzmin, lo, charge,
-                                    n_rz_azimuthal_modes, cost,
-                                    load_balance_costs_update_algo);
-    }
-    WARPX_PROFILE_VAR_STOP(blp_ppc_chd);
-
-#ifndef AMREX_USE_GPU
-    // CPU, tiling: atomicAdd local_rho into rho
-    WARPX_PROFILE_VAR_START(blp_accumulate);
-    (*rho)[pti].atomicAdd(local_rho, tb, tb, 0, icomp*nc, nc);
-    WARPX_PROFILE_VAR_STOP(blp_accumulate);
-#endif
-}
-
 #endif // CHARGEDEPOSITION_H_
