+++ Source/Particles/WarpXParticleContainer.cpp
@@ -9,6 +9,7 @@
  */
 #include "WarpXParticleContainer.H"
 
+#include "ABLASTR/ABLASTR.H"
 #include "Deposition/ChargeDeposition.H"
 #include "Deposition/CurrentDeposition.H"
 #include "Pusher/GetAndSetPosition.H"
@@ -647,11 +648,11 @@ WarpXParticleContainer::DepositCharge (WarpXParIter& pti, RealVector& wp,
         amrex::LayoutData<amrex::Real>* costs = WarpX::getCosts(lev);
         amrex::Real* cost = costs ? &((*costs)[pti.index()]) : nullptr;
 
-        DepositChargeDoIt<WarpXParticleContainer>
-            (pti, wp, ion_lev, rho, icomp, offset, np_to_depose,
+        ABLASTR::DepositCharge<WarpXParticleContainer>
+            (pti, wp, ion_lev, rho, icomp, nc, offset, np_to_depose,
              local_rho[thread_num], lev, depos_lev, this->charge,
-             WarpX::nox, WarpX::noy, WarpX::noz, ng_rho, dx, ref_ratio, nc,
-             xyzmin, cost, WarpX::n_rz_azimuthal_modes, WarpX::load_balance_costs_update_algo);
+             WarpX::nox, WarpX::noy, WarpX::noz, ng_rho, dx, xyzmin, ref_ratio,
+             cost, WarpX::n_rz_azimuthal_modes, WarpX::load_balance_costs_update_algo);
     }
 }
